requirements.txt
streamlit
pandas<2.2.0
numpy<2.0.0
# llama-cpp-python

requests
beautifulsoup4
feedparser
selenium
webdriver-manager

chromadb>=0.4.0
sentence-transformers

lxml
readability-lxml

python-dotenv
yfinance

# FinanceDataReader
finance-datareader
tabulate